+++ pages/[category].vue
@@ -1,0 +1,16 @@
+<template>
+  <BlogList :active-tag="category" />
+</template>
+
+<script>
+import BlogList from '~/components/BlogList.vue'
+
+export default {
+  components: { BlogList },
+  computed: {
+    category() {
+      return this.$route.params.category
+    }
+  }
+}
+</script>
